Fix hotspot overflow
