+++ components/image.tsx
@@ -1,9 +1,8 @@
 import clsx from 'classnames';
-import Link from 'next/link';
 import { Image as CrystallizeImage } from '@crystallize/reactjs-components';
 
 import { ImageVariant, Showcase } from '@/generated/graphql';
-import { Price } from './price';
+import { HotSpot } from './hot-spot';
 
 type CrystallizeImageVariants = React.ComponentProps<typeof CrystallizeImage>['variants'];
 
@@ -12,6 +11,7 @@ type ImageProps = {
     preserveRatio?: boolean;
     showShowcases?: boolean;
     className?: string;
+    wrapperClassName?: string;
     variants?: Array<Omit<ImageVariant, 'size' | 'key'> | null> | null;
     ogVariants?: Array<Omit<ImageVariant, 'size' | 'key'> | null> | null;
     sizes?: string;
@@ -31,6 +31,7 @@ export const Image = ({
     loading,
     ogVariants,
     sizes,
+    wrapperClassName,
     ...image
 }: ImageProps) => {
     const styles = {
@@ -48,13 +49,11 @@ export const Image = ({
     return (
         <div
             style={styles}
-            className={clsx(orientation, { 'crystallize-image': !preserveRatio, relative: showShowcases }, className)}
+            className={clsx(orientation, 'w-full h-full', wrapperClassName, {
+                'crystallize-image': !preserveRatio,
+                relative: showShowcases,
+            })}
         >
-            {showShowcases && (
-                <div className="absolute top-0 left-0 w-full h-full">
-                    {image?.showcases?.map((showcase, index) => <HotSpot key={index} showcase={showcase} />)}
-                </div>
-            )}
             <CrystallizeImage
                 {...image}
                 url={image.url ?? undefined}
@@ -62,71 +61,13 @@ export const Image = ({
                 alt={altText ?? undefined}
                 loading={loading}
                 sizes={sizes ?? '(max-width: 640px) 500w, 768w'}
+                className={className}
             />
-        </div>
-    );
-};
-
-export const HotSpot = ({ showcase }: { showcase: any }) => {
-    if (!showcase?.hotspot?.x) return null;
-    const x = showcase?.hotspot.x * 100;
-    const y = showcase?.hotspot.y * 100;
-    const product = showcase?.items?.[0] || showcase?.variants?.[0];
-    const image = product?.firstImage || product?.defaultVariant?.firstImage;
-    const price = product?.defaultPrice || product?.defaultVariant?.defaultPrice;
-    const link = product?.product?.path || product?.path;
-
-    const HotspotWrapper = ({ children }: { children: React.ReactElement[] }) => {
-        const classes =
-            'group absolute w-8 h-8 flex items-center justify-center rounded-full bg-dark/40 -translate-x-1/2 -translate-y-1/2';
-        const styles = {
-            left: `${x}%`,
-            top: `${y}%`,
-        };
-        if (!link)
-            return (
-                <div className={classes} style={styles}>
-                    {children}
+            {showShowcases && (
+                <div className="absolute top-0 left-0 w-full h-full">
+                    {image?.showcases?.map((showcase, index) => <HotSpot key={index} showcase={showcase} />)}
                 </div>
-            );
-        return (
-            <Link href={link} className={classes} style={styles}>
-                {children}
-            </Link>
-        );
-    };
-    return (
-        <HotspotWrapper>
-            <div className="bg-light w-4 h-4 rounded-full z-10"></div>
-            <div
-                className={clsx(
-                    'group-hover:w-auto group-hover:h-auto top-0 absolute w-0 h-0 overflow-hidden bg-dark rounded-lg transition-all',
-                    {
-                        'left-0 pl-8': x <= 50,
-                        'right-0 pr-8 pl-2': x > 50,
-                    },
-                )}
-            >
-                {product && (
-                    <div className=" gap-4 flex w-full text-light py-2 pr-6 text-sm items-center">
-                        <div className="w-12 h-12 overflow-hidden rounded-sm shrink-0 [&_img]:object-cover [&_img]:object-center bg-light">
-                            <CrystallizeImage
-                                {...image}
-                                alt={image?.altText}
-                                className="h-full w-12"
-                                loading={'lazy'}
-                                sizes="100px"
-                            />
-                        </div>
-                        <div className="flex flex-col overflow-hidden gap-1">
-                            <span className="text-ellipsis text-nowrap w-full">{product.name}</span>
-                            <span className="text-xs font-medium">
-                                <Price price={price} />
-                            </span>
-                        </div>
-                    </div>
-                )}
-            </div>
-        </HotspotWrapper>
+            )}
+        </div>
     );
 };
